perf(app): defer dashboard overview client graph
