+++ apps/app/src/app/(private-routes)/dashboard/overview/DashboardOverviewRouteBoundary.tsx
@@ -1,0 +1,14 @@
+'use client'
+
+import dynamic from 'next/dynamic'
+
+import RouteLoading from '@nl/ui/custom/route-loading'
+
+const DashboardOverviewClient = dynamic(() => import('./DashboardOverviewClient'), {
+  ssr: false,
+  loading: () => <RouteLoading label="Loading dashboard overview" />,
+})
+
+export default function DashboardOverviewRouteBoundary(): React.ReactNode {
+  return <DashboardOverviewClient />
+}
